merge_all float to int for iteration

+++ src/merge_all.py
@@ -12,7 +12,7 @@ outfile=sys.argv[2]
 
 dk_full_df= None
 
-replicates=((len(sys.argv)-3)/2)
+replicates=int((len(sys.argv)-3)/2)
 for i in range(3,(len(sys.argv)-replicates-1)):
     file=sys.argv[i]
     rep=sys.argv[i+replicates]
